modularizar
